Sales Endpoint Implementation - update completed. waiting for IPlayerDAO

+++ SoccerProject/src/main/java/com/itacademy/soccer/dto/Sale.java
@@ -60,7 +60,7 @@ public class Sale {
 		this.limitDate = limitDate;
 	}
 
-	public double getInitialPrice() {
+	public float getInitialPrice() {
 		return initialPrice;
 	}
